==> cell_line_processing/02_Tahoe/notebooks/00_GCSFileSystem.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import gcsfs

fs = gcsfs.GCSFileSystem()

gcp_base_path = "gs://arc-ctc-tahoe100/"

fs.ls(gcp_base_path)

fs.ls('gs://arc-ctc-tahoe100/2025-02-25/metadata/')
