==> cell_line_processing/02_Tahoe/notebooks/01_tahoe_save_metadata.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd
from scipy.sparse import csr_matrix
import anndata
import pandas as pd
import pubchempy as pcp
from tqdm import tqdm
from datasets import load_dataset
import pandas as pd

df = pd.read_parquet('../../NO_BACKUP/tahoe/obs_metadata_filtered.parquet')

# Load dataset in streaming mode
#ds = 
drug_meta = load_dataset("tahoebio/Tahoe-100m", 'drug_metadata', split="train").to_pandas()
# View the first record
#next(ds.iter(1))

drug_meta.to_parquet('../../NO_BACKUP/tahoe/drug_meta.parquet', index=False)

# Load dataset in streaming mode
#ds = 
gene_meta = load_dataset("tahoebio/Tahoe-100m", 'gene_metadata', split="train").to_pandas()
# View the first record
#next(ds.iter(1))

gene_meta.to_parquet('../../NO_BACKUP/tahoe/gene_meta.parquet', index=False)

cell_line_metadata = load_dataset("vevotx/Tahoe-100M","cell_line_metadata", split="train").to_pandas()
cell_line_metadata.head()

cell_line_metadata.to_parquet('../../NO_BACKUP/tahoe/cell_line_info.parquet', index=False)
