==> cell_line_processing/02_Tahoe/notebooks/02_tahoe_lincs_common_genes.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd

df_meta = pd.read_parquet('../../NO_BACKUP/tahoe/obs_metadata_filtered.parquet')

df_drugs = pd.read_parquet('../../NO_BACKUP/tahoe/drug_meta.parquet')

df_drugs['drug'].unique()

df_lincs = pd.read_parquet('../../NO_BACKUP/lincs_2020/signature_response_features_r2_top0.7_final.parquet')

df_meta[~df_meta['cell_name'].isin(set(df_lincs['cellosaurus_id']))]

set(df_meta['cell_name']).intersection(set(df_lincs['cellosaurus_id']))

set(df_meta['drug'].str.replace('-', '').str.lower()).intersection(set(df_lincs['cmap_name'].str.replace('-', '').str.lower()))

len({'afatinib',
 'altretamine',
 'anastrozole',
 'artesunate',
 'azd8055',
 'belinostat',
 'bexarotene',
 'bi78d3',
 'bortezomib',
 'busulfan',
 'celecoxib',
 'cepharanthine',
 'ciclopirox',
 'cinacalcet',
 'clofarabine',
 'clotrimazole',
 'cobimetinib',
 'decitabine',
 'digitoxin',
 'diphenhydramine',
 'docetaxel',
 'drospirenone',
 'eplerenone',
 'everolimus',
 'finasteride',
 'fulvestrant',
 'gefitinib',
 'glasdegib',
 'gsk1059615',
 'hexylresorcinol',
 'hydroxyurea',
 'idoxuridine',
 'irinotecan',
 'ixazomib',
 'lonafarnib',
 'ly2090314',
 'mebendazole',
 'menadione',
 'methotrexate',
 'mifepristone',
 'mozavaptan',
 'neratinib',
 'nevirapine',
 'nvpbhg712',
 'olaparib',
 'omeprazole',
 'oxaprozin',
 'paclitaxel',
 'palbociclib',
 'panobinostat',
 'penfluridol',
 'pimozide',
 'ponatinib',
 'raltitrexed',
 'resveratrol',
 'ritonavir',
 'sunitinib',
 'tak733',
 'terfenadine',
 'trametinib',
 'tranilast',
 'triclosan',
 'trifluridine',
 'valdecoxib',
 'vandetanib',
 'vemurafenib',
 'vismodegib'})

df_gene = pd.read_parquet('../../NO_BACKUP/tahoe/gene_meta.parquet')

df_lincs_genes = pd.read_parquet('../../NO_BACKUP/lincs_2020/genes_info.parquet')

df_gene

df_lincs_genes

gene_sym = df_lincs_genes[df_lincs_genes['gene_symbol'].isin(df_gene['gene_symbol'])]['gene_id']

df_lincs_genes.merge(gene_sym, on='gene_id').to_parquet('../../NO_BACKUP/lincs_2020/common_genes_lincs_tahoe.parquet', index=False)

df_lincs_genes[~df_lincs_genes['gene_symbol'].isin(df_gene['gene_symbol'])][df_lincs_genes[~df_lincs_genes['gene_symbol'].isin(df_gene['gene_symbol'])]['ensembl_id'].isin(df_lincs_genes[df_lincs_genes['gene_symbol'].isin(df_gene['gene_symbol'])]['ensembl_id'])]
