==> cell_line_processing/02_Tahoe/notebooks/03_tahoe_pharmacodb_common_entities.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd

df_cell = pd.read_parquet('../../NO_BACKUP/pharmacodb/cell_lines.parquet')

df_cell

df_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/compounds_pharmdb.parquet')

df_comp

df = pd.read_parquet('../../NO_BACKUP/tahoe/metadata/obs_metadata.parquet')

df

pd.read_csv('../../NO_BACKUP/tahoe/metadata/obs_metadata_descriptions.tsv', sep='\t').iloc[15]['description']

tahoe_celllines = df['cell_line'].unique()

tahoe_celllines

len(set(tahoe_celllines).intersection(set(df_cell['accession_id'])))

df_drugs = pd.read_parquet('../../NO_BACKUP/tahoe/drug_meta.parquet')

df_drugs = df_drugs[~df_drugs['pubchem_cid'].isna()].reset_index(drop=True)

df_drugs['pubchem_cid'] = df_drugs['pubchem_cid'].astype(int).astype(str)

df_drugs[df_drugs['pubchem_cid'].isin(df_comp['pubchem'])]['drug'].unique()

df = df[df['drug'].isin(df_drugs[df_drugs['pubchem_cid'].isin(df_comp['pubchem'])]['drug'].unique())]

df = df[df['cell_line'].isin(df_cell['accession_id'].unique())]

df = df[df['pass_filter'] == 'full']

df.to_parquet('../../NO_BACKUP/tahoe/obs_metadata_filtered.parquet', index=False)

pd.DataFrame(df['cell_line'].drop_duplicates()).to_parquet('../../NO_BACKUP/tahoe/cell_lines_id.parquet', index=False)
