==> cell_line_processing/02_Tahoe/notebooks/06_process_output_expressions.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd
import anndata

### Download all the expression

df1 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate1_agg_z_scores.h5ad')

df3 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate3_agg_z_scores.h5ad')

df4 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate4_agg_z_scores.h5ad')

df5 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate5_agg_z_scores.h5ad')

df7 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate7_agg_z_scores.h5ad')

df9 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate9_agg_z_scores.h5ad')

df10 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate10_agg_z_scores.h5ad')

df11 = anndata.read_h5ad('../../NO_BACKUP/tahoe/aggregated/plate11_agg_z_scores.h5ad')

df = pd.concat([df1.obs, df3.obs, df4.obs, df5.obs, df7.obs, df9.obs, df10.obs, df11.obs])

df.reset_index(drop=True).to_parquet('../../NO_BACKUP/tahoe/computed_signaturs_meta.parquet', index=False)

pd.concat([pd.DataFrame(df1.layers['mean']), 
           pd.DataFrame(df3.layers['mean']),
           pd.DataFrame(df4.layers['mean']),
           pd.DataFrame(df5.layers['mean']),
           pd.DataFrame(df7.layers['mean']),
           pd.DataFrame(df9.layers['mean']),
           pd.DataFrame(df10.layers['mean']),
           pd.DataFrame(df11.layers['mean']),
           ]).to_parquet('../../NO_BACKUP/tahoe/sig_matrix.parquet', index=False)

df1.var.reset_index().to_parquet('../../NO_BACKUP/tahoe/sig_genes.parquet')
