==> cell_line_processing/02_Tahoe/notebooks/07_tahoe_fingerprints.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')
# https://github.com/UnixJunkie/mol2ecfp4/blob/master/ecfp4.py

import pandas as pd

from rdkit import Chem
from rdkit.Chem import AllChem

import pandas as pd
import numpy as np

### Filter drugs

df_drugs = pd.read_parquet('../../NO_BACKUP/tahoe/drug_meta.parquet')

df_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/compounds_pharmdb.parquet')

df_comp

df_drugs = df_drugs[~df_drugs['pubchem_cid'].isna()].reset_index(drop=True)

df_drugs['pubchem_cid'] = df_drugs['pubchem_cid'].astype(int).astype(str)

comp_overlap = df_drugs[df_drugs['pubchem_cid'].isin(df_comp['pubchem'])].reset_index(drop=True)

from rdkit.Chem import rdFingerprintGenerator
fp_diameter = 4
fp_radius = fp_diameter / 2
rdkgen = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=2048)
fps = []
cfps = []

for smile in comp_overlap['canonical_smiles'].values:
    if not smile is None:
        mol = Chem.MolFromSmiles(smile)
        fp = rdkgen.GetFingerprintAsNumPy(mol)
        cfp = rdkgen.GetCountFingerprintAsNumPy(mol)
        fps.append(fp)
        cfps.append(cfp)
    else:
        fps.append(None)
        cfps.append(None)

df_fps = pd.DataFrame(np.array(fps))

rename_fps_col = {}
for i in range(2048):
    rename_fps_col[i] = 'fps_' + str(i)

df_fps = df_fps.rename(columns=rename_fps_col)

df_fps['pubchem_cid'] = comp_overlap['pubchem_cid']

comp_overlap

df_fps

df_fps['drug'] = comp_overlap['drug']

df_fps['smiles'] = comp_overlap['canonical_smiles']

df_fps.merge()

df_cfps = pd.DataFrame(np.array(cfps))

rename_cfps_col = {}
for i in range(2048):
    rename_cfps_col[i] = 'cfps_' + str(i)
df_cfps = df_cfps.rename(columns=rename_cfps_col)

df_fps['drug'] = comp_overlap['drug']
df_fps['smiles'] = comp_overlap['canonical_smiles']
df_cfps = pd.DataFrame(np.array(cfps))

rename_cfps_col = {}
for i in range(2048):
    rename_cfps_col[i] = 'cfps_' + str(i)
df_cfps = df_cfps.rename(columns=rename_cfps_col)

df_fps.to_parquet('../../NO_BACKUP/tahoe/comp_overlap_fps.parquet', index=False)
df_cfps.to_parquet('../../NO_BACKUP/tahoe/comp_overlap_cfps.parquet', index=False)
