==> cell_line_processing/02_Tahoe/notebooks/08_tahoe_pharmacodb_overlapping.ipynb <==
import sys
from tqdm import tqdm as tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd
import anndata

df_sig = pd.read_parquet('../../NO_BACKUP/tahoe/computed_signaturs_meta.parquet')

df_resp = pd.read_parquet('../../NO_BACKUP/pharmacodb/responses.parquet')

df_drugs = pd.read_parquet('../../NO_BACKUP/tahoe/drug_meta.parquet')

df_drugs = df_drugs[~df_drugs['pubchem_cid'].isna()].reset_index(drop=True)
df_drugs['pubchem_cid'] = df_drugs['pubchem_cid'].astype(int).astype(str)

pharm_cell_lines = pd.read_parquet('../../NO_BACKUP/pharmacodb/cell_lines.parquet')

pharm_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/compounds_pharmdb.parquet')

pharm_comp = pharm_comp.drop_duplicates(subset=['inchikey'])[['id', 'inchikey', 'pubchem']]

pharm_comp = pharm_comp.dropna()

pharm_comp

df_resp_ = df_resp.copy()

### overlapping

df_resp = df_resp_.merge(pharm_cell_lines[['id', 'accession_id']].dropna(), left_on='cell_line_id', right_on='id', how='left').rename(columns={'id_x': 'comp_id', 'idy':'cell_id'})

df_resp = df_resp.merge(pharm_comp, left_on='comp_id', right_on='id', how='left')

df_resp = df_resp.dropna(subset=['pubchem']).reset_index(drop=True)

df_sig['cell_line'] = df_sig['cl_sm_dose'].str.split('_').str[:2].str.join('_')

df_sig['drug'] = df_sig['cl_sm_dose'].str.split('_').str[2:].str[:-1].str.join('_')

df_sig['dose'] = df_sig['cl_sm_dose'].str.split('_').str[2:].str[-1]

df_sig = df_sig.merge(df_drugs[['drug', 'pubchem_cid']], on='drug', how='left').dropna()

df_sig['cell_line_drug'] = df_sig['cell_line'] + '_' + df_sig['pubchem_cid']

df_resp['cell_line_drug'] = df_resp['accession_id'] + '_' + df_resp['pubchem']

df_resp = df_resp.dropna(subset=['cell_line_drug']).reset_index(drop=True)

df_resp

df_resp_overlap = df_resp[df_resp['cell_line_drug'].isin(df_sig['cell_line_drug'].unique())]

df_resp_overlap.to_parquet('../../NO_BACKUP/tahoe/responses_overlap.parquet', index=False)

df_sig_overlap = df_sig[df_sig['cell_line_drug'].isin(df_resp['cell_line_drug'].unique())]

df_sig_overlap.to_parquet('../../NO_BACKUP/tahoe/sig_overlap.parquet', index=False)
