==> cell_line_processing/02_Tahoe/notebooks/11_combine_all_the_data_inferred.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')
import pandas as pd
import numpy as np

### mapping genes

matrix = pd.read_parquet('../../NO_BACKUP/tahoe/sig_matrix.parquet')

df_lincs = pd.read_parquet('../../NO_BACKUP/lincs_2020/signature_response_features_r2_top0.7_final.parquet')

df_lincs = df_lincs.iloc[0]

df_sig_genes = pd.read_parquet('../../NO_BACKUP/tahoe/sig_genes.parquet')

df_sig_genes

df_sig_genes = df_sig_genes.reset_index().rename(columns={'index': 'tahoe_gene_id', 'gene_name': 'gene_symbol'})

df_sig_genes['tahoe_gene_id'] = df_sig_genes['tahoe_gene_id'].astype(int).astype(str)

lincs_genes = pd.read_parquet('../../NO_BACKUP/lincs_2020/genes_info.parquet')

lincs_genes = lincs_genes[['gene_id', 'gene_symbol', 'ensembl_id', 'gene_title']].rename(columns={'gene_id': 'lincs_gene_id'})

lincs_genes['lincs_gene_id'] = lincs_genes['lincs_gene_id'].astype(int).astype(str)

lincs_genes = lincs_genes.merge(df_sig_genes, on='gene_symbol', how='left')

lincs_genes

lincs_genes_order = df_lincs.columns[df_lincs.columns.isin(lincs_genes['lincs_gene_id'])]

matrix_t = matrix.T.reset_index()

matrix_t = matrix_t.rename(columns={'index': 'tahoe_gene_id'})

matrix_t['tahoe_gene_id'] = matrix_t['tahoe_gene_id'].astype(str)

genes = lincs_genes.merge(matrix_t, on='tahoe_gene_id', how='left')

genes = genes.drop(columns=['gene_symbol', 'ensembl_id', 'gene_title', 'tahoe_gene_id'])

genes = genes.set_index('lincs_gene_id')

genes = genes.loc[lincs_genes_order]

#genes.reset_index().to_parquet('../../NO_BACKUP/tahoe/genes_tahoe_lincs.parquet', index=False)

matrix_mapped = genes.T

matrix_mapped

# map to sig

df_sig = pd.read_parquet('../../NO_BACKUP/tahoe/computed_signaturs_meta.parquet')

df_sig

df = pd.concat([df_sig, matrix_mapped], axis=1)

df

df_sig_resp = pd.read_parquet('../../NO_BACKUP/tahoe/signature_response.parquet')

df_sig_resp = df_sig_resp.drop_duplicates('cl_sm_dose')

df = df.drop_duplicates('cl_sm_dose')

df_sig_resp

df_comp = pd.read_parquet('../../NO_BACKUP/tahoe/comp_overlap_fps.parquet')

df_comp = df_comp.drop(columns='drug')
df_comp = df_comp.drop_duplicates('pubchem_cid')

column_to_move = df_comp.pop("pubchem_cid")
df_comp.insert(0, "pubchem_cid", column_to_move)
column_to_move = df_comp.pop("smiles")
df_comp.insert(1, "smiles", column_to_move)


df_sig_resp_ = df_sig_resp.merge(df_comp, on='pubchem_cid', how='left')

df_sig_resp_

df_comp

df_sig_resp_final = df_sig_resp_.merge(df, on='cl_sm_dose', how='left')

pharm_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/compounds_pharmdb.parquet')

pharm_comp = pharm_comp[['inchikey', 'pubchem']].dropna()

df_lincs = df_lincs.merge(pharm_comp, left_on='inchi_key', right_on='inchikey')
#[['cell_drug']].drop_duplicates('cell_drug')

df_cell_drug = df_lincs['cellosaurus_id'] + '_' + df_lincs['pubchem']

lincs_cells = df_lincs['cellosaurus_id'].unique()

df_sig_resp_final[df_sig_resp_final['cell_line'].isin(lincs_cells)].to_parquet('../../NO_BACKUP/tahoe/final_tahoe_overlapping_lincs.parquet', index=False)

df_sig_resp_final[~df_sig_resp_final['cell_line'].isin(lincs_cells)].to_parquet('../../NO_BACKUP/tahoe/final_tahoe_no_overlappings.parquet', index=False)

df = pd.read_parquet('../../NO_BACKUP/tahoe/final_tahoe_no_overlappings.parquet')

df_lincs = pd.DataFrame(df_lincs).T.drop(columns=['sig_id', 'pert_id', 'cellosaurus_id','inchi_key','cell_drug','nearest_dose','responses','cmap_name', 'smiles'])

df = df.drop(columns=['index', 'cl_sm_dose', 'cell_line', 'drug', 'dose', 'pubchem_cid', 'cell_line_drug', 'responses','smiles'])

pd.concat([df, df_lincs])
